refactor: package arrangement (make domain package) [#34]
