--- spring/src/main/java/com/bancow/bancowback/common/exception/BizException.java
@@ -1,7 +1,0 @@
-package com.bancow.bancowback.common.exception;
-
-public class BizException extends RuntimeException {
-    public BizException(String message) {
-        super(message);
-    }
-}
