--- spring/src/main/java/com/bancow/bancowback/common/util/mail/entity/MailTemplate.java
@@ -1,36 +1,0 @@
-package com.bancow.bancowback.common.util.mail.entity;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class MailTemplate {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	private String templateId;
-
-	private String title;
-	private String contents;
-
-	private String sendEmail;
-	private String sendUserName;
-
-	private LocalDateTime regDate;
-}
