--- spring/src/main/java/com/bancow/bancowback/common/util/token/entity/Token.java
@@ -1,40 +1,0 @@
-package com.bancow.bancowback.common.util.token.entity;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-import com.bancow.bancowback.manager.entity.Manager;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Token {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column
-	private String token;
-
-	@OneToOne(fetch = FetchType.LAZY)
-	private Manager manager;
-
-	@Column(name = "expired_date")
-	private LocalDateTime expiredDate;
-}
