+++ spring/src/main/java/com/bancow/bancowback/domain/common/dto/Response.java
@@ -1,0 +1,15 @@
+package com.bancow.bancowback.domain.common.dto;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class Response<T> {
+	private T data;
+	private HttpStatus status;
+}
