+++ spring/src/main/java/com/bancow/bancowback/domain/common/exception/BizException.java
@@ -1,0 +1,7 @@
+package com.bancow.bancowback.domain.common.exception;
+
+public class BizException extends RuntimeException {
+    public BizException(String message) {
+        super(message);
+    }
+}
