+++ spring/src/main/java/com/bancow/bancowback/domain/manager/entity/Manager.java
@@ -1,0 +1,50 @@
+package com.bancow.bancowback.domain.manager.entity;
+
+import java.time.LocalDateTime;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Manager {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column
+	private String email;
+
+	@Column
+	private String password;
+
+	@Column
+	private String username;
+
+	@Enumerated(EnumType.STRING)
+	@Column(name = "manager_status")
+	private ManagerStatus managerStatus;
+
+	@Column(name = "create_date")
+	private LocalDateTime createDate;
+
+	@Column(name = "update_date")
+	private LocalDateTime updateDate;
+
+}
