+++ spring/src/main/java/com/bancow/bancowback/domain/manager/entity/ManagerStatus.java
@@ -1,0 +1,5 @@
+package com.bancow.bancowback.domain.manager.entity;
+
+public enum ManagerStatus {
+	SUPER, ADMIN, PENDING_EMAIL, PENDING_SUPER;
+}
