--- spring/src/main/java/com/bancow/bancowback/manager/entity/ManagerStatus.java
@@ -1,5 +1,0 @@
-package com.bancow.bancowback.manager.entity;
-
-public enum ManagerStatus {
-	SUPER, ADMIN, PENDING_EMAIL, PENDING_SUPER;
-}
